+++ src/main/java/ru/t1/java/demo/aop/LogAspect.java
@@ -38,13 +38,12 @@ public class LogAspect {
     @AfterReturning(
             pointcut = "@annotation(ru.t1.java.demo.aop.annotations.HandlingResult)",
             returning = "result")
-    public void handleResult(JoinPoint joinPoint, List<Client> result) {
+    public void handleResult(JoinPoint joinPoint, List<Object> result) {
         log.info("В результате выполнения метода {}", joinPoint.getSignature().toShortString());
-//        log.info("получен результат: {} ", result);
+        log.info("получен результат: {} ", result);
         log.info("Подробности: \n");
-
-        result = isNull(result) ? List.of() : result;
-
+        List<Object> result_ = isNull(result) ? List.of() : result;
+        log.info(result_.toString());
     }
 
 }
