+++ src/main/java/ru/t1/java/demo/config/KafkaConsumerGenerator.java
@@ -1,0 +1,45 @@
+package ru.t1.java.demo.config;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
+import org.springframework.kafka.core.*;
+import org.springframework.kafka.listener.CommonErrorHandler;
+import org.springframework.kafka.listener.ContainerProperties;
+import org.springframework.kafka.listener.DefaultErrorHandler;
+import org.springframework.util.backoff.FixedBackOff;
+
+
+@Slf4j
+public class KafkaConsumerGenerator {
+    public static <T> ConcurrentKafkaListenerContainerFactory<String, T> buildKafkaListenerContainerFactory(
+            ConsumerFactory<String, T> consumerFactory) {
+        ConcurrentKafkaListenerContainerFactory<String, T> factory
+                = new ConcurrentKafkaListenerContainerFactory<>();
+        factoryBuilder(consumerFactory, factory);
+        return factory;
+    }
+
+    private static <T> void factoryBuilder(ConsumerFactory<String, T> consumerFactory,
+                                    ConcurrentKafkaListenerContainerFactory<String, T> factory) {
+        factory.setConsumerFactory(consumerFactory);
+        factory.setBatchListener(true);
+        factory.setConcurrency(1);
+        factory.getContainerProperties().setAckMode(ContainerProperties.AckMode.MANUAL_IMMEDIATE);
+        factory.getContainerProperties().setPollTimeout(5000);
+        factory.getContainerProperties().setMicrometerEnabled(true);
+        factory.setCommonErrorHandler(errorHandler());
+    }
+
+    private static CommonErrorHandler errorHandler() {
+        DefaultErrorHandler handler = new DefaultErrorHandler(new FixedBackOff(1000, 3));
+        handler.addNotRetryableExceptions(IllegalStateException.class);
+        handler.setRetryListeners((record, ex, deliveryAttempt) ->
+                log.error(" RetryListeners message = {}, offset = {} deliveryAttempt = {}",
+                        ex.getMessage(), record.offset(), deliveryAttempt));
+        return handler;
+    }
+
+
+
+
+}
