+++ src/main/java/ru/t1/java/demo/config/KafkaProducerFactoryConfig.java
@@ -1,0 +1,22 @@
+package ru.t1.java.demo.config;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.core.*;
+import java.util.HashMap;
+import java.util.Map;
+
+@Slf4j
+@Configuration
+@RequiredArgsConstructor
+public class KafkaProducerFactoryConfig {
+    private final KafkaCommonProperties kafkaCommonProperties;
+
+    @Bean
+    public <T> ProducerFactory<String, T> producerFactory() {
+        Map<String, Object> producerProperties = new HashMap<>(kafkaCommonProperties.buildProducerCommonProperties());
+        return new DefaultKafkaProducerFactory<>(producerProperties);
+    }
+}
