+++ src/main/java/ru/t1/java/demo/exception/AccountException.java
@@ -1,0 +1,14 @@
+package ru.t1.java.demo.exception;
+
+public class AccountException extends RuntimeException {
+    public AccountException() {
+    }
+
+    public AccountException(String message) {
+        super(message);
+    }
+
+    public AccountException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
