+++ src/main/java/ru/t1/java/demo/repository/AccountRepository.java
@@ -16,20 +16,14 @@ import java.util.Optional;
 @Repository
 public interface AccountRepository extends JpaRepository<Account, Long> {
     @Query("select id from Account a where a.accountId = ?1")
-    Long findIdByAccountId(@NonNull String accountId);
+    Optional<Long> findIdByAccountId(@NonNull String accountId);
 
     @Query("select accountId from Account a where a.id = ?1")
-    String findAccountIdById(@NonNull Long id);
-
-    @Query("select status from Account a where a.id = ?1")
-    AccountStatus findStatusById(@NonNull Long id);
-
-    @Query("select balance from Account a where a.id = ?1")
-    BigDecimal findBalanceById(@NonNull Long id);
+    Optional<String> findAccountIdById(@NonNull Long id);
 
     @Transactional
     @Modifying
-    @Query("update Account a set a.balance = ?2 where a.id = ?1")
-    int updateBalanceById(@NonNull Long id, BigDecimal updatedBalance);
+    @Query("update Account a set a.balance = ?2 where a.accountId = ?1")
+    int updateBalanceById(@NonNull String id, BigDecimal updatedBalance);
 
 }
