+++ src/main/java/ru/t1/java/demo/repository/ClientRepository.java
@@ -6,11 +6,13 @@ import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Repository;
 import ru.t1.java.demo.model.Client;
 
+import java.util.Optional;
+
 @Repository
 public interface ClientRepository extends JpaRepository<Client, Long> {
     @Query("select id from Client c where c.clientId = ?1")
-    Long findIdByClientId(@NonNull String clientId);
+    Optional<Long> findIdByClientId(@NonNull String clientId);
 
     @Query("select clientId from Client c where c.id = ?1")
-    String findClientIdById(@NonNull Long id);
+    Optional<String> findClientIdById(@NonNull Long id);
 }
